+++ packages/dev-tools/src/features/index.ts
@@ -1,0 +1,3 @@
+export { print } from "./console";
+
+export type { status } from "./console";