+++ packages/dev-tools/src/main.ts
@@ -1,1 +1,3 @@
 export * from "./configs";
+export * from "./features";
+export * from "./plugins";